Simplify ios_runner by reusing ssh_runner

+++ runners/ios_runner.py
@@ -5,20 +5,22 @@ import os, sys, re
 import time
 from waflib.TaskGen import feature, after_method
 from waflib import Utils, Task, Logs, Options
-from basic_runner import BasicRunner
+from ssh_runner import SSHRunner
 
-class IOSRunner(BasicRunner):
-    def save_result(self, results, usbmux_proc):
+class IOSRunner(SSHRunner):
+
+    def save_result(self, results, ssh_cmd):
+
         # Kill the usbmux process
-        usbmux_proc.kill()
+        if self.usbmux_proc != None:
+            usbmux_proc.kill()
 
-        super(IOSRunner, self).save_result(results)
+        super(IOSRunner, self).save_result(results, ssh_cmd)
 
     def run(self):
 
         bld = self.generator.bld
-
-        results = []
+        self.usbmux_proc = None
 
         dest_dir = '/private/var/mobile/tmp'
         localport = '22222'
@@ -33,7 +35,7 @@ class IOSRunner(BasicRunner):
                 stderr=Utils.subprocess.PIPE,
                 stdout=Utils.subprocess.PIPE)
             # Wait for a second so that the process can start
-            time.sleep(1.0)
+            time.sleep(2.0)
 
             return proc
 
@@ -46,84 +48,7 @@ class IOSRunner(BasicRunner):
         ssh_cmd = ['ssh', '-t', '-p', localport, ssh_target]
 
         # Start the usbmux daemon to forward 'localport' to port 22 on USB
-        usbmux_proc = start_proc(usbmux_cmd)
+        self.usbmux_proc = start_proc(usbmux_cmd)
 
-        # Enumerate the test files
-        file_list = [test_input.abspath() for test_input in self.test_inputs]
-
-        # Add the binary
-        binary = self.inputs[0]
-        file_list.append(binary.abspath())
-
-        # Copy all files in file_list
-        result = self.run_cmd(scp_cmd + file_list + [ssh_target+':'+dest_dir])
-        results.append(result)
-
-        if result['return_code'] != 0:
-            self.save_result(results, usbmux_proc)
-            return
-
-        run_binary_cmd = "./{0}".format(binary)
-
-        # if this is a benchmark and we need to retrieve the result file
-        if  bld.has_tool_option('run_benchmark') and \
-            bld.has_tool_option('python_result'):
-            # Add the benchmark python result output filename option
-            run_binary_cmd += " --pyfile={}".format(
-                bld.get_tool_option("python_result"))
-
-        # Add the given run command modifications
-        run_binary_cmd = self.format_command(run_binary_cmd)
-
-        # Echo the exit code after the shell command
-        result = self.run_cmd(
-            ssh_cmd + \
-            ["cd {0};{1};echo shellexit:$?".format(dest_dir, run_binary_cmd)])
-
-        results.append(result)
-
-        if result['return_code'] != 0:
-            self.save_result(results, usbmux_proc)
-            return
-
-        # Almost done. Look for the exit code in the output
-        # and fail if nonzero
-        match = re.search('shellexit:(\d+)', result['stdout'])
-
-        if not match:
-            result =  {'cmd': 'Looking for shell exit', 'return_code': -1,
-                       'stdout': '', 'stderr': 'Failed to find exitcode'}
-
-            results.append(result)
-            self.save_result(results, usbmux_proc)
-            return
-
-        if match.group(1) != "0":
-            result =  {'cmd': 'Shell exit indicates error',
-                       'return_code': match.group(1),
-                       'stdout': '',
-                       'stderr': 'Exit code was %s' % match.group(1)}
-
-            results.append(result)
-            self.save_result(results, usbmux_proc)
-            return
-
-        # Everything seems to be fine, lets pull the output file if needed
-        if  bld.has_tool_option('run_benchmark') and \
-            bld.has_tool_option('python_result'):
-            output_file = bld.get_tool_option("python_result")
-
-            # Remove the old benchmark if it exists
-            self.run_cmd(["rm", "-f", output_file])
-
-            benchmark_result = os.path.join(dest_dir,output_file)
-
-            result = self.run_cmd(
-                scp_cmd + ['{0}:{1}'.format(ssh_target, benchmark_result), '.'])
-            results.append(result)
-
-            if result['return_code'] != 0:
-                self.save_result(results, usbmux_proc)
-                return
-
-        self.save_result(results, usbmux_proc)
+        # Call the 'run_ssh' method of the SSHRunner class
+        super(IOSRunner, self).run_ssh(ssh_cmd, scp_cmd, ssh_target, dest_dir)
